feat: move auth logic in useAuth
