+++ packages/client/src/pages/Forum/Forum.tsx
@@ -1,10 +1,9 @@
-import React, { ChangeEventHandler, useMemo, useState, useEffect } from 'react'
+import React, { ChangeEventHandler, useMemo, useState } from 'react'
 import { Wrapper } from '../../components/Wrapper'
 import { CreationBlockContent } from './components/CreationBlockContent'
 import { Topic } from './components/Topic'
 import { mockTopics } from './mock'
-import { useAppSelector } from '../../store/hooks'
-import { useNavigate } from 'react-router-dom'
+import { useAuth } from '../../hooks/useAuth'
 
 import classes from './Forum.module.scss'
 
@@ -14,12 +13,8 @@ const plugTopic = {
 }
 
 const Forum: React.FC = function () {
-  const user = useAppSelector(state => state.user)
-  const navigate = useNavigate()
-
-  useEffect(() => {
-    if (!user) navigate('/login')
-  }, [user])
+  const [checkAuth] = useAuth()
+  checkAuth('private')
 
   const [currentTopics, setTopics] = useState(mockTopics)
   const [isTopicCretionOn, setTopicCreationState] = useState(false)
