+++ packages/client/src/pages/Game/Game.tsx
@@ -1,15 +1,11 @@
-import React, { useEffect } from 'react'
+import React from 'react'
 import { Game as GameComponent } from '../../components/Game'
-import { authController } from '../../controllers/AuthController'
-import { useNavigate } from 'react-router-dom'
 import { Button } from '../../components/Button'
-import { useAppSelector, useAppDispatch } from '../../store/hooks'
-import { deleteUser } from '../../store/slices/userSlice'
+import { useAuth } from '../../hooks/useAuth'
 
 const Game: React.FC = function () {
-  const navigate = useNavigate()
-  const dispatch = useAppDispatch()
-  const user = useAppSelector(state => state.user)
+  const [checkAuth, auth] = useAuth()
+  checkAuth('private')
 
   document.addEventListener(
     'keydown',
@@ -21,15 +17,6 @@ const Game: React.FC = function () {
     false
   )
 
-  function onExitClick() {
-    authController.logout().then(res => {
-      if (res) {
-        dispatch(deleteUser())
-        navigate('/login')
-      }
-    })
-  }
-
   function onFullscreenClick() {
     if (!document.fullscreenElement) {
       document.documentElement.requestFullscreen()
@@ -38,14 +25,10 @@ const Game: React.FC = function () {
     }
   }
 
-  useEffect(() => {
-    if (!user) navigate('/login')
-  }, [user])
-
   return (
     <div>
       <GameComponent />
-      <Button onClick={onExitClick}>Выйти</Button>
+      <Button onClick={auth.exit}>Выйти</Button>
       <Button onClick={onFullscreenClick}>Fullscreen</Button>
     </div>
   )
